requirements.txt
# Phase 1 — Feature extraction
librosa>=0.10.0
numpy>=1.24.0
pandas>=2.0.0
soundfile>=0.12.0
scikit-learn>=1.4.0
joblib>=1.3.0
xgboost>=2.0.0
python-dotenv>=1.0.0
